feat(cli)!: rename devire command to checkout
