+++ src/commands.rs
@@ -1,4 +1,4 @@
 pub mod add;
+pub mod checkout;
 pub mod commit;
-pub mod derive;
 pub mod init;
